Extract Crest from Program
